0412 Parked vehicles api updated

+++ src/main/java/com/parking/repository/ParkedVehicleRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ParkedVehicleRepository extends JpaRepository<ParkedVehicle, Integer> {
+public interface ParkedVehicleRepository extends JpaRepository<ParkedVehicle, Long> {
 
     ParkedVehicle findByLicensePlate(String licensePlate);
