+++ src/main/java/com/parking/repository/RouteCycleRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RouteCycleRepository extends JpaRepository<RouteCycle, Integer> {
+public interface RouteCycleRepository extends JpaRepository<RouteCycle, Long> {
     RouteCycle findByCycleNumber(Integer cycleNumber);
 
     @Query(value = "SELECT MAX(cycle_number) FROM route_cycle", nativeQuery = true)
